Extract Linux-specific code to a separate module
Linux was the only target platform considered when Yafut was originally
designed.  However, as it gained the ability to operate on Yaffs file
system images using POSIX APIs, it became potentially useful on other
operating systems as well (e.g. macOS).  The obstacle preventing that is
the unconditional use of Linux-specific APIs (MTD ioctls and the related
structures) in the source code.  Extract all Linux-specific code to a
separate module that only gets compiled on that system and make existing
Linux-specific modules conditionally-compiled as well.  Implement a
generic counterpart of the Linux-specific module to be used on other
operating systems (where only Yaffs file system images will be
supported).

Suggested-by: Mark Mentovai <mark@mentovai.com>
