+++ CMakeLists.txt
@@ -68,21 +68,35 @@ add_executable(
 	src/file.c
 	src/file_driver_posix.c
 	src/file_driver_yaffs.c
-	src/ioctl.c
 	src/layout.c
 	src/log.c
 	src/main.c
 	src/options.c
 	src/storage.c
 	src/storage_driver_image.c
-	src/storage_driver_nand.c
-	src/storage_driver_nor.c
 	src/util.c
 	src/yaffs_glue.c
 	src/ydriver.c
-	src/ydriver_ioctl.c
 	src/ydriver_posix.c
 )
+
+if(LINUX)
+	target_sources(
+		yafut
+		PRIVATE
+		src/ioctl.c
+		src/storage_driver_nand.c
+		src/storage_driver_nor.c
+		src/storage_platform_linux.c
+		src/ydriver_ioctl.c
+	)
+else()
+	target_sources(
+		yafut
+		PRIVATE
+		src/storage_platform_generic.c
+	)
+endif()
 
 set_target_properties(
 	yafut
