+++ src/file.h
@@ -8,7 +8,6 @@
 
 enum file_type {
 	FILE_TYPE_POSIX,
-	FILE_TYPE_MTD,
 	FILE_TYPE_YAFFS,
 };
